==> code/01_SIRatio.qmd <==
---
title: 'Figure 1: SI Ratio & Time in IZ Zone'
format:
    html:
        code-fold: false
        self-contained: true
        toc: true
editor:
    render-on-save: false
---

Figure 1 demonstrates that social interaction ratio is a sensitive metric of stress effect in defeat males but not witness females. This code generates Fig 1b and Fig 1c as well as statistics for the results section **SI ratio identifies marginal stress effects in witness females**.

### Libraries & Config

```{r}
#| label: libraries
#| message: false
#| code-fold: true

library(here)
library(glue)
library(dplyr)
library(tidyr)
library(knitr)
library(pander)
library(ggplot2)
library(ggpubr)
library(pwr)
library(effectsize)
library(grid)
library(cowplot)
library(emmeans)
```

```{r}
#| label: config + paths + data
#| code-fold: true

source(here('styles.R'))
data_path <- here('data')
SIT_diff <- readRDS(glue('{data_path}/SIT_diff_primary.rds'))
SIT_raw <- readRDS(glue('{data_path}/SIT_raw_primary.rds'))
```

------------------------------------------------------------------------

### SI Ratio effects are larger in males than in females (Fig. 1b)

Following exposure to CSW/DS, SI ratio is reduced in males and females compared to non-stressed controls. This effect is driven by changes in stressed animals, as male and female control mice do not differ significantly from each other.

```{r}
#| label: Fig 1b prep
#| code-fold: true

# PLOT ----

point <- ggplot(SIT_diff, aes(Condition, IZ_time, color = full_cond, fill = full_cond)) + 
              geom_hline(yintercept = 1) + 
              geom_jitter(width = 0.2, alpha = 0.2, height = 0, size = 0.5) +
              stat_summary(fun = 'mean', geom = 'point', size = 0.5) + 
              stat_summary(fun.data = 'mean_se', geom = 'errorbar', 
                           linewidth = 1, width = 0.4) + 
              scale_y_continuous(name = 'SI Ratio', expand = c(0,0,0.3,0)) + 
              facet_wrap(Sex ~ ., nrow = 2) + 
              scale_color_manual(values = col_full) + 
              scale_fill_manual(values = col_full) + 
              plot_theme + 
              theme(legend.position = 'None',
                    axis.title.x = element_blank(),
                    plot.margin = unit(c(0,0,0,0), 'cm'))

hist <- ggplot(SIT_diff, aes(y=IZ_time, color = full_cond, fill = full_cond)) + 
              geom_hline(yintercept = 1) + 
              geom_density(alpha = 0.4) +
              scale_x_continuous(expand = c(0,0), limits = c(0,1.2),
                                 breaks = c(0,0.5,1),  name = c('Density')) + 
              scale_y_continuous(name = 'SI Ratio', expand = c(0,0,0.3,0)) + 
              scale_color_manual(values = col_full) + 
              scale_fill_manual(values = col_full) + 
              plot_theme + 
              facet_wrap(Sex ~ ., nrow = 2) + 
              theme(legend.position = 'None',
                    axis.line = element_blank(),
                    axis.ticks = element_blank(),
                    axis.title = element_blank(),
                    axis.text = element_blank(),
                    strip.text = element_blank(),
                    plot.margin = unit(c(0,0,0,0), 'cm'))

# STATS ----

# ANOVA
aov_SIratio <- aov(IZ_time ~ Condition * Sex + cohort, SIT_diff) 
sum_aov_SIratio <- summary(aov_SIratio)

sidak_cond_sex <- pairs(emmeans(aov_SIratio, ~ Condition | Sex), adjust = 'sidak')
sidak_sex_cond <- pairs(emmeans(aov_SIratio, ~ Sex | Condition), adjust = 'sidak')

```

::: {layout="[28, -2, 70]"}
::: {plot}
```{r}
#| label: Fig 1b plot
#| fig-width: 3
#| fig-asp: 2.1
#| fig-align: center
#| echo: false

plot_grid(point, hist + clean_theme(), 
          ncol = 2, nrow = 1, align = 'hv', rel_widths = c(1.5,1))
```
:::

::: {table}
```{r}
#| label: Fig 1b stats
#| echo: false

pander(sum_aov_SIratio)
kable(rbind(sidak_cond_sex, sidak_sex_cond), 
            digits = 2, align='c', 
            caption = 'Sidak Post-Hoc Planned Contrasts')
```
:::
:::

We then calculated effect sizes for males and females. For this, we performed a t-test for male and female data separately as a basis to calculate hedge's *g*. This value was then used to calculate the sample size required **per group** for an adequately powered experiment (*power* = 0.8), as well as to calculate the power currently achieved when using sample sizes ranging from 8-15 animals per group.

```{r}
#| label: effect sizes
#| code-fold: true

# ES
m_data <- SIT_diff[SIT_diff$Sex == 'Male',]
t_SIratio_m <- t.test(IZ_time ~ Condition, data = m_data) # t-test
g_t_SIratio_m <- hedges_g(IZ_time ~ Condition, data = m_data)$Hedges_g # hedge's g
n_t_m <- pwr.t.test(d = g_t_SIratio_m, power = 0.8,
                    sig.level = 0.05, type = "two.sample") # sample size for power = 0.8
pwr_t_low_m <- pwr.t.test(d = g_t_SIratio_m, n = 8,
                          sig.level = 0.05, type = "two.sample") # power with n = 8
pwr_t_high_m <- pwr.t.test(d = g_t_SIratio_m, n = 15,
                           sig.level = 0.05, type = "two.sample") # power with n = 15

f_data <- SIT_diff[SIT_diff$Sex == 'Female',]
t_SIratio_f <- t.test(IZ_time ~ Condition, data = f_data) # t-test
g_t_SIratio_f <- hedges_g(IZ_time ~ Condition, data = f_data)$Hedges_g # hedge's g
n_t_f <- pwr.t.test(d = g_t_SIratio_f, power = 0.8,
                    sig.level = 0.05, type = "two.sample") # sample size for power = 0.8
pwr_t_low_f <- pwr.t.test(d = g_t_SIratio_f, n = 8,
                          sig.level = 0.05, type = "two.sample") # power with n = 8
pwr_t_high_f <- pwr.t.test(d = g_t_SIratio_f, n = 15,
                           sig.level = 0.05, type = "two.sample") # power with n = 15

es <- tibble("Sex" = c('Male', 'Female'),
             "t Value" = c(round(t_SIratio_m$statistic, 2), round(t_SIratio_f$statistic, 2)),
             "df" = c(round(t_SIratio_m$parameter, 2), round(t_SIratio_f$parameter, 2)),
             "Hedge's g" = c(round(g_t_SIratio_m, 2), round(g_t_SIratio_f, 2)),
             "n (pwr = 0.8)" = c(ceiling(n_t_m$n), ceiling(n_t_f$n)),
             "pwr (n = 8)" = c(round(pwr_t_low_m$power, 2), round(pwr_t_low_f$power, 2)),
             "pwr (n = 15)" = c(round(pwr_t_high_m$power, 2), round(pwr_t_high_f$power, 2)))

# Variance test
var_ctl <- var.test(IZ_time ~ Sex, SIT_diff[SIT_diff$Condition == 'Ctl',], 
                    alternative = "two.sided")
var_stress <- var.test(IZ_time ~ Sex, SIT_diff[SIT_diff$Condition == 'Stress',], 
                       alternative = "two.sided")
var_test <- tibble("Group" = c('Control', 'Stress'),
                   "F Value" = c(var_ctl$statistic, var_stress$statistic),
                   "df" = c(paste0(var_ctl$parameter[1], ',', var_ctl$parameter[2]), 
                            paste0(var_stress$parameter[1], ',', var_stress$parameter[2])),
                   "p Value" = c(var_ctl$p.value, var_stress$p.value))

kable(es)
kable(var_test)
```

### Raw data confirm differential effects in defeated males and witness females (Fig. 1c)

Finally, we tested if female witnesses showed evidence of stress effects in the raw time in the interaction zone, that may be masked when computing the SI ratio. Defeated males show reduced time in their interaction zone only during the Target trial. In contrast, witness females show a modest increase in time in the interaction zone in the No Target trial, and a modest decrease during the Target trial compared to controls.

```{r}
#| label: Fig 1c prep
#| message: false
#| code-fold: true

# PLOT ----
point_raw <- ggplot(SIT_raw, aes(Condition, IZ_time, color = full_cond, fill = full_cond)) + 
                        geom_jitter(width = 0.2, alpha = 0.2, height = 0, size = 0.5) +
                        stat_summary(fun = 'mean', geom = 'point', size = 0.5) + 
                        stat_summary(fun.data = 'mean_se', geom = 'errorbar', 
                                    linewidth = 1, width = 0.4) + 
                        scale_y_continuous(name = 'Time in Interaction Zone (s)', expand = c(0,0,0.3,0)) +
                        facet_grid(Sex ~ Trial_type) + 
                        scale_color_manual(values = col_full) + 
                        scale_fill_manual(values = col_full) + 
                        plot_theme + 
                        theme(legend.position = 'None',
                                axis.title.x = element_blank(), 
                                panel.spacing.y = unit(2, "lines"))

# STATS ----

# ANOVAs per sex
m_data <- SIT_raw[SIT_raw$Sex == 'Male',]
m_aov <- aov(IZ_time ~ Condition * Trial_type + cohort + Error(id_full), m_data)
f_data <- SIT_raw[SIT_raw$Sex == 'Female',]
f_aov <- aov(IZ_time ~ Condition * Trial_type + cohort + Error(id_full), f_data)

# Sidak 
sidak_cond_tt_m <- pairs(emmeans(m_aov, ~ Condition | Trial_type), adjust = 'sidak')
sidak_tt_cond_m <- pairs(emmeans(m_aov, ~ Trial_type | Condition), adjust = 'sidak')
sidak_cond_tt_f <- pairs(emmeans(f_aov, ~ Condition | Trial_type), adjust = 'sidak')
sidak_tt_cond_f <- pairs(emmeans(f_aov, ~ Trial_type | Condition), adjust = 'sidak')

```

```{r}
#| label: Fig 1c plot and stats
#| message: false
#| echo: false
#| fig-width: 5
#| fig-asp: 1.25
#| fig-align: center

point_raw

pander(summary(m_aov), caption = 'rmANOVA in Males')
kable(rbind(sidak_cond_tt_m,
            sidak_tt_cond_m), digits = 5, align='c', caption = 'Sidak Post-Hoc Planned Contrasts in Males')

pander(summary(f_aov), caption = 'rmANOVA in Females')
kable(rbind(sidak_cond_tt_f,
            sidak_tt_cond_f), digits = 5, align='c', caption = 'Sidak Post-Hoc Planned Contrasts in Females')

```

------------------------------------------------------------------------

### Session Info

```{r}
#| label: Session Info
#| echo: False
sessionInfo()
```
