==> code/03_PCA.qmd <==
---
title: 'Figure 3: PCA; Changes in Up- and Downregulated syllables'
format:
    html:
        code-fold: false
        self-contained: true
        toc: true
editor:
    render-on-save: false
---

Figure 3 demonstrates that female witness mice show behavioral variability distinct of that observed in males as a result of CSW/DS, and that this behavioral variability is driven by an attenuated behavioral regulation in response to target presence. This code generates Fig 3 and Fig. S3 as well as statistics for the results section **Dimensionality reduction identifies distinct sources of stress-related variance in defeat males and witness females**.

### Libraries & Config

```{r}
#| label: libraries
#| message: false
#| code-fold: true

library(here)
library(glue)
library(dplyr)
library(tidyr)
library(knitr)
library(magrittr)
library(pander)
library(ggplot2)
library(ggpubr)
library(ggpmisc)
library(ggrepel)
library(cowplot)
library(emmeans)
```

```{r}
#| label: config + paths + data
#| code-fold: true

source(here('styles.R'))
data_path <- here('data')
SIT_diff <- readRDS(glue('{data_path}/SIT_diff_primary.rds'))
SIT_raw <- readRDS(glue('{data_path}/SIT_raw_primary.rds'))
SIT_z <- readRDS(glue('{data_path}/SIT_z_primary.rds'))
syl_info <- tibble(read.csv(glue('{data_path}/syllable_info.csv')))
```

### Run PCA

We first ran a PCA on SI ratio, corner ratio and the NT - T difference for each syllable. We looked exclusively at the first two principle components.

```{r}
#| label: pca
#| code-fold: true
#| warning: false

# Include IZ Ratio, Corner Ratio, and all syllables
SIT_pca <- SIT_diff[,c(grep('_time', names(SIT_diff)),
                       grep('syl', names(SIT_diff)))]
pca <- prcomp(SIT_pca, scale = TRUE, center = TRUE) #run PCA

# Add PCA results to full data frame
SIT_diff$PC1 <- pca$x[,1]
SIT_diff$PC2 <- pca$x[,2]

# Get factor loadings and prep for plotting
SIT_pca <- tibble(var = names(SIT_pca), 
                  PC1 = as.numeric(pca$rotation[,1]), 
                  PC2 = as.numeric(pca$rotation[,2]))
SIT_pca$new_label <- syl_info$new_label[match(SIT_pca$var, syl_info$metric)]
SIT_pca$avg_freq <- syl_info$avg_freq[match(SIT_pca$var, syl_info$metric)]
SIT_pca$new_label <- ifelse(is.na(SIT_pca$new_label), 
                            gsub('_time', ' Ratio', SIT_pca$var), 
                            SIT_pca$new_label)
SIT_pca$new_label[SIT_pca$new_label == 'IZ Ratio'] = 'SI Ratio'
SIT_pca$social_identity <- ifelse(SIT_pca$new_label %in% c('SI Ratio', 'Corner Ratio'), 
                                  'A', substr(SIT_pca$new_label, 1, 1))

# Calcualte PC1 and PC2 mean for plotting
SIT_diff %>%
  group_by(Sex, Condition, full_cond) %>%
  summarise(avg_PC1 = mean(PC1), avg_PC2 = mean(PC2),
            sd_PC1 = sd(PC1), sd_PC2 = sd(PC2)) -> SIT_pca_mean

```

------------------------------------------------------------------------

### PCA Plots (Fig. 3a, 3b)

The PC plot shows that CSW/DS induces behavioral variability largely captured along PC1 in defeat males, but along PC2 in witness females. This analysis not only confirms that male and female mice differ in their behavioral response after CSW/DS in the SIT, but also emphasizes that CSW/DS indeed causes behavioral changes in witness females, disputing the hypothesis that the stress model does not work.

```{r}
#| label: Fig 3a prep and plot
#| fig-width: 10
#| fig-asp: 0.5
#| fig-align: center
#| code-fold: true
#| warning: false
#| error: false

# Female plots
F_pca <- ggplot(SIT_diff[SIT_diff$Sex == 'Female',]) + 
           geom_hline(yintercept = 0) + 
           geom_vline(xintercept = 0) + 
           geom_point(aes(PC1, PC2, color=full_cond), 
                          size = 2, alpha = 0.3) + 
           geom_point(data = SIT_pca_mean[SIT_pca_mean$Sex == 'Female',], 
                      aes(avg_PC1, avg_PC2, 
                          color=full_cond), size = 5) +
           ylim(c(-4,4)) + xlim(c(-4,4)) + 
           scale_color_manual(values = col_full[c(1,2)]) + 
           plot_theme + 
           theme(legend.position = 'none',
                 plot.margin = unit(c(0,0,0,0), 'cm'))

F_hist_pc1 <- ggplot(SIT_diff[SIT_diff$Sex == 'Female',], 
                     aes(PC1, color=full_cond, fill = full_cond)) + 
                geom_hline(yintercept = 0) +
                geom_density(alpha = 0.5) + 
                xlim(c(-4,4)) + 
                scale_color_manual(values = col_full[c(1,2)]) + 
                scale_fill_manual(values = col_full[c(1,2)]) + 
                plot_theme + 
                theme(legend.position = 'none',
                      plot.margin = unit(c(0,0,0,0), 'cm')) + 
                clean_theme()

F_hist_pc2 <- ggplot(SIT_diff[SIT_diff$Sex == 'Female',], 
                     aes(PC2, color=full_cond, fill = full_cond)) + 
                geom_hline(yintercept = 0) +
                geom_density(alpha = 0.5) + 
                xlim(c(-4,4)) + 
                scale_color_manual(values = col_full[c(1,2)]) + 
                scale_fill_manual(values = col_full[c(1,2)]) + 
                coord_flip() +
                plot_theme + 
                theme(legend.position = 'none',
                      plot.margin = unit(c(0,0,0,0), 'cm')) + 
                clean_theme()

F_plot <- plot_grid(F_hist_pc1, NULL, F_pca, F_hist_pc2, align = 'hv', 
                    rel_widths = c(3,1), rel_heights = c(1,3))

# Male plots
M_pca <- ggplot(SIT_diff[SIT_diff$Sex == 'Male',]) + 
           geom_hline(yintercept = 0) + 
           geom_vline(xintercept = 0) + 
           geom_point(aes(PC1, PC2, color=full_cond), 
                          size = 2, alpha = 0.3) + 
           geom_point(data = SIT_pca_mean[SIT_pca_mean$Sex == 'Male',], 
                      aes(avg_PC1, avg_PC2, 
                          color=full_cond), size = 5) +
           ylim(c(-4,4)) + xlim(c(-4,4)) + 
           scale_color_manual(values = col_full[c(3,4)]) + 
           plot_theme + 
           theme(legend.position = 'none',
                 plot.margin = unit(c(0,0,0,0), 'cm'))

M_hist_pc1 <- ggplot(SIT_diff[SIT_diff$Sex == 'Male',], 
                     aes(PC1, color=full_cond, fill = full_cond)) + 
                geom_hline(yintercept = 0) +
                geom_density(alpha = 0.5) + 
                xlim(c(-4,4)) + 
                scale_color_manual(values = col_full[c(3,4)]) + 
                scale_fill_manual(values = col_full[c(3,4)]) + 
                plot_theme + 
                theme(legend.position = 'none',
                      plot.margin = unit(c(0,0,0,0), 'cm')) + 
                clean_theme()

M_hist_pc2 <- ggplot(SIT_diff[SIT_diff$Sex == 'Male',], 
                     aes(PC2, color=full_cond, fill = full_cond)) + 
                geom_hline(yintercept = 0) +
                geom_density(alpha = 0.5) + 
                xlim(c(-4,4)) + 
                scale_color_manual(values = col_full[c(3,4)]) + 
                scale_fill_manual(values = col_full[c(3,4)]) + 
                coord_flip() +
                plot_theme + 
                theme(legend.position = 'none',
                      plot.margin = unit(c(0,0,0,0), 'cm')) + 
                clean_theme()

M_plot <- plot_grid(M_hist_pc1, NULL, M_pca, M_hist_pc2, align = 'hv', 
                    rel_widths = c(3,1), rel_heights = c(1,3))

# Plot 
plot_grid(M_plot, F_plot)
```

The factor loadings of the PC plot show that PC1 is mostly driven by location-based metrics, i.e., SI ratio and corner ratio. In contrast, PC2 is driven by syllable frequency differences. Notably, Up- and Downregulated syllables have opposing effects on PC2.

```{r}
#| label: Fig 3b prep and plot
#| code-fold: true
#| fig-width: 5
#| fig-asp: 1
#| fig-align: center
#| warning: false
#| error: false

ggplot(SIT_pca, aes(PC1, PC2, color = social_identity, fill = social_identity)) + 
  geom_hline(yintercept = 0) + geom_vline(xintercept = 0) + 
  geom_label_repel(aes(label = new_label), 
                   size = 5, alpha = 0.2, seed = 1234 , fontface = 'bold')+ 
  geom_label_repel(aes(label = new_label), 
                   size = 5, alpha = 1, seed = 1234,  
                   label.size = NA, fill = NA, fontface = 'bold') + 
  scale_color_manual(values = col_syls) + 
  scale_fill_manual(values = col_syls) + 
  ylim(c(-0.6,0.6)) + xlim(c(-0.8,0.8)) + 
  plot_theme + 
  clean_theme() + theme(legend.position = 'none')
```

### Importance of Syllable Social Identity (Fig. 3c, 3d)

Given that Up- and Downregulated syllables split along the axis of PC2, which in turn differentiates control and witness females, we next looked at the average changes in syllable frequency observed within each social identity group. We show that Upregulated syllables are more upregulated in defeated males compared to controls. In witness females, we see an attenuated regulation in both Up- and Downregulated syllables.

We summarize these observations by quantifying the average absolute syllable frequency change, which we term the Social Modulation Score (SMS). We observed a significant increase in SMS in males, and a significant decrease in witness females.

```{r}
#| label: Fig 3cd prep and plot
#| code-fold: true
#| fig-width: 10
#| fig-asp: 0.66
#| fig-align: center
#| warning: false
#| error: false

# Annotate syllables with social identity
N_syls <- syl_info$metric[grepl('N',syl_info$new_label)]
U_syls <- syl_info$metric[grepl('U',syl_info$new_label)] 
D_syls <- syl_info$metric[grepl('D',syl_info$new_label)]

SIT_z$Social <- factor(ifelse(SIT_z$metric %in% N_syls, 'No Change', 
                              ifelse(SIT_z$metric %in% U_syls, 'Up', 'Down')),
                              levels = c('Up', 'No Change', 'Down'))

# Calculate mean per social identity type and abs change
SIT_z %>%
  group_by(id_full, full_cond, Sex, Condition, Social, cohort) %>%
  summarise(Social_avg = mean(z_diff)) -> SIT_syl_avg_freq_z_diff

SIT_z %>%
  mutate(all = 'All Syllables') %>% # So that plot can be titled by facet
  group_by(id_full, full_cond, Sex, Condition, cohort, all) %>%
  summarise(SMS = mean(abs(z_diff))) -> SIT_avg_freq_z_diff

# Plot for each social identity type
### Not faceted because of different y axes
U_plot <- ggplot(SIT_syl_avg_freq_z_diff[SIT_syl_avg_freq_z_diff$Social == 'Up',], 
                 aes(Condition, Social_avg, color = full_cond, fill = full_cond)) + 
            geom_hline(yintercept = 0) + 
            geom_jitter(width = 0.2, alpha = 0.2, height = 0, size = 0.3) +
            stat_summary(fun.y = 'mean', geom = 'point', size = 2) + 
            stat_summary(fun.data = 'mean_se', geom = 'errorbar', 
                         linewidth = 1, width = 0.4) + 
            scale_y_continuous(name = 'Mean Syllable Frequency Diff. (Z Score)', 
                               limits = c(-10,10), breaks = c(-10,0,10)) + 
            facet_grid(Sex ~ Social, scale = 'free_y') + 
            scale_color_manual(values = col_full) + 
            scale_fill_manual(values = col_full) + 
            plot_theme + 
            theme(legend.position = 'None',
                  axis.title.x = element_blank())

N_plot <- ggplot(SIT_syl_avg_freq_z_diff[SIT_syl_avg_freq_z_diff$Social == 'No Change',], 
                 aes(Condition, Social_avg, color = full_cond, fill = full_cond)) + 
            geom_hline(yintercept = 0) + 
            geom_jitter(width = 0.2, alpha = 0.2, height = 0, size = 0.3) +
            stat_summary(fun.y = 'mean', geom = 'point', size = 2) + 
            stat_summary(fun.data = 'mean_se', geom = 'errorbar', 
                         linewidth = 1, width = 0.4) + 
            scale_y_continuous(name = 'Mean Syllable Frequency Diff. (Z Score)', 
                     limits = c(-5,5), breaks = c(-5,0,5)) + 
            facet_grid(Sex ~ Social, scale = 'free_y') + 
            scale_color_manual(values = col_full) + 
            scale_fill_manual(values = col_full) + 
            plot_theme + 
            theme(legend.position = 'None',
                  axis.title.x = element_blank())

D_plot <- ggplot(SIT_syl_avg_freq_z_diff[SIT_syl_avg_freq_z_diff$Social == 'Down',], 
                 aes(Condition, Social_avg, color = full_cond, fill = full_cond)) + 
            geom_hline(yintercept = 0) + 
            geom_jitter(width = 0.2, alpha = 0.2, height = 0, size = 0.3) +
            stat_summary(fun.y = 'mean', geom = 'point', size = 2) + 
            stat_summary(fun.data = 'mean_se', geom = 'errorbar', 
                         linewidth = 1, width = 0.4) + 
            scale_y_continuous(name = 'Mean Syllable Frequency Diff. (Z Score)', 
                     limits = c(-2,2), breaks = c(-2,0,2)) + 
            facet_grid(Sex ~ Social, scale = 'free_y') + 
            scale_color_manual(values = col_full) + 
            scale_fill_manual(values = col_full) + 
            plot_theme + 
            theme(legend.position = 'None',
                  axis.title.x = element_blank())

SMS_plot <- ggplot(SIT_avg_freq_z_diff, 
                   aes(Condition, SMS, color = full_cond, fill = full_cond)) + 
              geom_jitter(width = 0.2, alpha = 0.2, height = 0, size = 0.3) +
              stat_summary(fun.y = 'mean', geom = 'point', size = 2) + 
              stat_summary(fun.data = 'mean_se', geom = 'errorbar', 
                           linewidth = 1, width = 0.4) + 
              ylab('Social Modulation Score (SMS)') + 
              facet_grid(Sex ~ all) + 
              scale_color_manual(values = col_full) + 
              scale_fill_manual(values = col_full) + 
              plot_theme + 
              theme(legend.position = 'None',
                    axis.title.x = element_blank())

plot_grid(U_plot, N_plot, D_plot, SMS_plot, align = 'hv', nrow = 1)
```

```{r}
#| label: Fig 3cd stats
#| code-fold: true
#| warning: false

# rmANOVA by social type
aov_social_m <- aov(Social_avg ~ Condition * Social + cohort + Error(id_full), 
                    SIT_syl_avg_freq_z_diff[SIT_syl_avg_freq_z_diff$Sex == 'Male',])
aov_social_f <- aov(Social_avg ~ Condition * Social + cohort + Error(id_full),
                     SIT_syl_avg_freq_z_diff[SIT_syl_avg_freq_z_diff$Sex == 'Female',])

sidak_social_m <- pairs(emmeans(aov_social_m, 
                        ~ Condition | Social), adjust = 'sidak')
sidak_social_f <- pairs(emmeans(aov_social_f, 
                        ~ Condition | Social), adjust = 'sidak')

# ANOVA for SMS
aov_sms_m <- aov(SMS ~ Condition + cohort, 
                    SIT_avg_freq_z_diff[SIT_avg_freq_z_diff$Sex == 'Male',])
aov_sms_f <- aov(SMS ~ Condition + cohort,
                    SIT_avg_freq_z_diff[SIT_avg_freq_z_diff$Sex == 'Female',])
```

```{r}
#| label: show stats
#| echo: false

pander(aov_social_m,
       caption = 'rmANOVA: Social type * Condition on Frequency Difference in Males')
kable(sidak_social_m, 
      digits = 5, align='c', 
      caption = 'Sidak Post-Hoc Planned Contrasts ')
pander(aov_social_f,
       caption = 'rmANOVA: Social type * Condition on Frequency Difference in Females')
kable(sidak_social_f, 
      digits = 5, align='c', 
      caption = 'Sidak Post-Hoc Planned Contrasts')

pander(aov_sms_m,
       caption = 'ANOVA: Condition on SMS in Males')
pander(aov_sms_f,
       caption = 'ANOVA: Condition on SMS in Females')
```

### SI Ratio and SMS correlation to PCs (Fig. 3e)

Finally, we confirm that SI ratio correlates with PC1 and that SMS indeed captures the variability identified by PC2.

```{r}
#| label: Fig 3e prep + plot + save PCs&SMS to df
#| code-fold: true
#| fig-width: 10
#| fig-asp: 0.5
#| fig-align: center
#| warning: false
#| error: false

SIT_diff %<>%
  mutate(SMS = SIT_avg_freq_z_diff$SMS[match(id_full, 
                                             SIT_avg_freq_z_diff$id_full)])
saveRDS(SIT_diff, glue('{data_path}/SIT_diff_primary_pcsms.rds')) # save rds data to include pcs and sms for later figures

pc1 <- ggplot(SIT_diff, aes(PC1, IZ_time)) + 
         geom_point(aes(color=full_cond), size = 2, alpha = 0.3) + 
         stat_poly_line(col = 'black') +
         stat_poly_eq(size = 7, label.y = 'top', label.x = 'right') +
         scale_color_manual(values = col_full) + 
         xlab('PC1') + ylab('SI Ratio') +
         xlim(c(-4,4)) + 
         plot_theme + 
         theme(legend.position = 'none')

pc2 <- ggplot(SIT_diff, aes(PC2, SMS)) + 
         geom_point(aes(color=full_cond), size = 2, alpha = 0.3) + 
         stat_poly_line(col = 'black') +
         stat_poly_eq(size = 7, label.y = 'top', label.x = 'right') +
         scale_color_manual(values = col_full) + 
         xlab('PC2') + ylab('SMS') +
         xlim(c(-4,4)) + 
         plot_theme + 
         theme(legend.position = 'none')

plot_grid(pc1, pc2, ncol = 2, nrow = 1, align = 'hv')
```

------------------------------------------------------------------------

### Defeat male-specific behavior: U1 Upregulation (Fig. S3)

We observed that defeat ales show a higher upregulation of upregulated syllables only. When further investigating this observation we see that this is driven by an increase specific to U1 in defeated males compared to control males.

```{r}
#| label: Fig S2a prep + plot
#| code-fold: true
#| warning: false

# Select relevant data and annotate new syllable label
SIT_z$new_label <- syl_info$new_label[match(SIT_z$metric, syl_info$metric)]
SIT_U_m <- SIT_z[grepl('U', SIT_z$new_label) & SIT_z$Sex == 'Male',]

# Plot 
Um_plot <- ggplot(SIT_U_m, aes(Condition, z_diff, color = Condition)) + 
               stat_summary(fun = 'mean', geom = 'point', size = 2) + 
               stat_summary(fun.data = 'mean_se', geom = 'errorbar', 
                            linewidth = 1, width = 0.4) + 
               geom_jitter(width = 0.2, alpha = 0.2, height = 0) +
               scale_color_manual(values = col_full[c(3,4)]) + 
               facet_grid(. ~ new_label) + 
               ylab('Mean Syllable Frequency Difference') +
               ylim(-15, 5) + 
               plot_theme + 
               theme(legend.position = 'none',
                     axis.title.x = element_blank())

# Stats
aov_freq <- aov(z_diff ~ Condition * new_label + cohort + Error(id_full), SIT_U_m)
sidak_freq <- pairs(emmeans(aov_freq, ~ Condition | new_label, adjust = 'sidak'))
```

```{r}
#| label: Fig S2a show
#| echo: false
#| warning: false
#| fig-width: 10
#| fig-asp: 0.5
#| fig-align: center

Um_plot
pander(aov_freq, caption = 'rmANOVA: Condition * Syllable for Upregulated syllables in Males')
kable(sidak_freq, digits = 5, align='c', 
      caption = 'Sidak Post-Hoc Planned Contrasts')
```

We then explored whether there mice express this behavior preferentially at specific locations in the arena and whether it is correlated with SI ratio. Indeed, mice in an example cohort (E15) display this syllable specifically around the interaction zone and in the corners, suggesting that it is a target induced behavior. Finally, we see that in stressed males, the frequency of U1 is correlated to SI ratio, such that animals with a lower SI ratio display a higher frequency of syllable U1.

```{r}
#| label: Fig S2b1 plot
#| code-fold: true
#| fig-width: 5
#| fig-asp: 0.9
#| fig-align: center
#| warning: false

U1 <- readRDS(glue('{data_path}/Syls_descriptives_E15.rds'))[['21']] # THIS DATA IS ONLY ON OSF, NOT GITHUB > Please download separately.
ggplot(U1, aes(centroid_y, centroid_x)) + 
  geom_hex(aes(fill = stat(cut(count, breaks = c(0,25,50,75,100,125,150,175,Inf), 
               labels = F, right = T, include.lowest = T))), bins = 30) + 
  scale_fill_continuous(name = 'count') + 
  plot_theme + 
  theme(legend.position = 'right',
        legend.title = element_blank(),
        axis.title = element_blank(),
        axis.line = element_blank(),
        axis.text = element_blank(), 
        axis.ticks = element_blank())
```

```{r}
#| label: Fig S2b2 plot
#| code-fold: true
#| fig-width: 5
#| fig-asp: 0.9
#| fig-align: center
#| warning: false

for (syl in substr(syl_info$metric,4,5)) {
   syl_data <- readRDS(glue('{data_path}/Syls_descriptives_E15.rds'))[[syl]]
   ggplot(syl_data, aes(centroid_y, centroid_x)) + 
                  geom_hex() + 
                  scale_fill_continuous(name = 'count') + 
                  plot_theme + 
                  theme(legend.position = 'right',
                        legend.key.width = unit(0.4, 'cm'),
                        legend.key.height = unit(1, 'cm'),
                        legend.title = element_blank(),
                        axis.title = element_blank(),
                        axis.line = element_blank(),
                        axis.text = element_blank(), 
                        axis.ticks = element_blank())
   ggsave(glue('/home/bagotlab/global/heike/StressDLC/moseq/SIT_cropped_HSonly_revision/syllable{syl}_arena.pdf', width = 3500, height = 3000, units = 'px'))
}

```

```{r}
#| label: Fig S2c plot
#| code-fold: true
#| fig-width: 10
#| fig-asp: 0.5
#| fig-align: center
#| warning: false

SIT_raw$SI_ratio <- SIT_diff$IZ_time[match(SIT_raw$id_full, SIT_diff$id_full)] #add SI ratio to Raw df
SIT_raw_m <- filter(SIT_raw, Sex == 'Male')

# U1 = syl21
cor_plot <- ggplot(SIT_raw_m, aes(syl21, SI_ratio, color = Condition)) + 
                 geom_point(size = 2, alpha = 0.3) + 
                 stat_poly_line() +
                 stat_poly_eq(size = 7, label.y = 'top', label.x = 'right') +
                 scale_color_manual(values = col_full[c(3,4)]) + 
                 facet_wrap(. ~ Trial_type, scales = 'free_x', nrow = 1) +
                 xlab('Frequency') + ylab('SI Ratio') +
                 plot_theme + 
                 theme(legend.position = 'none')

# correlations
Ctl_NT <- cor.test(SIT_raw_m$SI_ratio[SIT_raw_m$Condition == 'Ctl' & SIT_raw_m$Trial_type == 'No Target'], 
                   SIT_raw_m$syl21[SIT_raw_m$Condition == 'Ctl' & SIT_raw_m$Trial_type == 'No Target'])
Str_NT <- cor.test(SIT_raw_m$SI_ratio[SIT_raw_m$Condition == 'Stress' & SIT_raw_m$Trial_type == 'No Target'], 
                  SIT_raw_m$syl21[SIT_raw_m$Condition == 'Stress' & SIT_raw_m$Trial_type == 'No Target'])
Ctl_T <- cor.test(SIT_raw_m$SI_ratio[SIT_raw_m$Condition == 'Ctl' & SIT_raw_m$Trial_type == 'Target'], 
                   SIT_raw_m$syl21[SIT_raw_m$Condition == 'Ctl' & SIT_raw_m$Trial_type == 'Target'])
Str_T <- cor.test(SIT_raw_m$SI_ratio[SIT_raw_m$Condition == 'Stress' & SIT_raw_m$Trial_type == 'Target'], 
                  SIT_raw_m$syl21[SIT_raw_m$Condition == 'Stress' & SIT_raw_m$Trial_type == 'Target'])

# show
cor_plot
kable(rbind(Ctl_NT, Str_NT, Ctl_T, Str_T)[,1:4], digits = 5,
      caption = 'Pearson Correlation tests')
```

------------------------------------------------------------------------

## Session Info

```{r}
#| label: Session Info
#| echo: False
sessionInfo()
```
